Cleaned up code + 1.13.1 fully included (mirrors 1.13.2)

+++ crates/default_protocol/src/version/v1_13/util.rs
@@ -13,7 +13,7 @@ pub fn easy_send<T: PacketEncode + Send + Sync + 'static>(connection: &mut Unbou
     }))
 }
 
-pub fn build_compacted_data_array<E: Iterator<Item=i32>>(bits_per_block: u8, elements: E) -> Vec<u64> {
+pub fn build_compacted_data_array<E: Iterator<Item=u64>>(bits_per_block: u8, elements: E) -> Vec<u64> {
     let long_count: u32 = (SECTION_WIDTH * SECTION_HEIGHT * SECTION_LENGTH * bits_per_block as u16) as u32 / i64::BITS;
     let mut compacted_data = Vec::with_capacity(long_count as usize);
     let mut current_long = 0u64;
@@ -23,16 +23,16 @@ pub fn build_compacted_data_array<E: Iterator<Item=i32>>(bits_per_block: u8, ele
     for element in elements {
         let bit_shift = pos * bits_per_block + offset;
         if bit_shift < (i64::BITS - bits_per_block as u32) as u8 {
-            current_long |= (element as u64) << bit_shift;
+            current_long |= element << bit_shift;
             pos += 1;
         } else {
             offset = bit_shift - (i64::BITS - bits_per_block as u32) as u8;
-            current_long |= (element as u64) << bit_shift;
+            current_long |= element << bit_shift;
             compacted_data.push(current_long);
             current_long = 0u64;
             if offset != 0 {
                 let diff = bits_per_block - offset;
-                current_long |= (element as u64) >> diff;
+                current_long |= element >> diff;
             }
             pos = 0;
         }
